According to DuBois 2009, PP186.
For example, the `score` table in the `sampdb` sample database contains a `student_id` column, which we use to relate score rows to students in the `student` table.When we created these tables in Chapter 1, we set up some explicit relationships between them. One of these was that we declared `score.student_id` to be a foreign key for the `student.student_id` column.That prevents a row from being entered into the `score` table unless its student_id value exists in the `student` table. In other words, the foreign key prevents entry of scores for non-existent students.